Use correct @Generated annotation when compiling with `--release 8` (#34)
`@javax.annotation.processing.Generated` is not available for compilation when passing
`--release 8` to the compiler,
but it is on the classpath. So, checking for its availability with `Class.forName()`,
like Jilt does currently, will return the wrong result.

The solution is to use the `Elements` object to find the `@Generated` class,
which respects the `--release 8` compiler argument,
and returns `null` for `@javax.annotation.processing.Generated`
when that option is set.